requirements.txt
aiodns==2.0.0
async-timeout==3.0.1
attrs==19.3.0
brotlipy==0.7.0
cachetools==4.0.0
cchardet==2.1.5
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-WTF==0.14.3
gevent==1.4.0
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.3.0
google-cloud-spanner==1.13.0
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpc-google-iam-v1==0.12.3
grpcio==1.26.0
grpcio-gcp==0.2.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
multidict==4.7.4
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.19
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
visitor==0.1.3
Werkzeug==1.0.0
WTForms==2.2.1
yarl==1.4.2
